requirements.txt
asgiref==3.11.0
Django==6.0
django-environ==0.12.0
django-jazzmin==3.0.1
djangorestframework==3.16.1
drf-yasg==1.21.11
environ==1.0
inflection==0.5.1
packaging==25.0
pillow==12.0.0
psycopg2==2.9.11
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.4
uritemplate==4.2.0